+++ docs/source/changelog.rst
@@ -19,6 +19,8 @@ Changed
 * Remove support for group of groups, a |group| take as input a |pipeline| or a message
 * |get_result_backend| now raise NoResultBackend if there is no |ResultBackend|
 * Merged PipelineResult and GroupResult into |CollectionResult|
+* |message_get_result| on forgotten results now returns None
+
 
 `0.6.0`_ -- 2018-11-23
 ----------------------